initiated registration==correct
