+++ src/Admin.java
@@ -1,5 +1,6 @@
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
+import java.util.Scanner;
 
 public class Admin extends User{
 
@@ -30,17 +31,27 @@ public class Admin extends User{
             return false;
         }
     }
-    // public String initiatePatientRegistration(String email){
-    //     try {
-    //         ProcessBuilder pb = new ProcessBuilder("bash", "user-manager.sh", "initiate_patient_registration",email);
-    //         pb.inheritIO();
-    //         Process process = pb.start();
-            
-    //         process.waitFor();
-    //     }
-    //     catch(Exception e){
-    //         e.printStackTrace();
-    //     }
-    // }
+    public static void initiatePatientRegistration() {
+        System.out.println("Enter patient's email:");
+        Scanner scanner = new Scanner(System.in);
+        String email = scanner.nextLine();
+
+        try {
+            ProcessBuilder pb = new ProcessBuilder("bash", "../bash/user-manager.sh", "initiate_patient_registration", email);
+            pb.redirectErrorStream(true);
+            Process process = pb.start();
+            BufferedReader reader = new BufferedReader(new InputStreamReader(process.getInputStream()));
+            String line;
+            StringBuilder output = new StringBuilder();
+            while ((line = reader.readLine()) != null) {
+                output.append(line);
+            }
+            process.waitFor();
+            System.out.println(output.toString());
+        } catch (Exception e) {
+            e.printStackTrace();
+            System.out.println("Error in execution");
+        }
+    }
 
 }
